requirements.txt
django
gunicorn
django-import-export
